% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRoonData.R
\docType{data}
\name{targets}
\alias{targets}
\title{Example target analyte list}
\format{A \code{data.frame} with column \code{name}, \code{mz} and \code{rt},
  specifying the analyte name, \emph{m/z} and retention time (in seconds),
  respectively.}
\usage{
targets
}
\description{
The \code{targets} dataset contains information for 23 target analytes and
can directly be used with the \code{\link[patRoon]{screenTargets}} function.
}
\keyword{datasets}
