% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRoonData.R
\docType{data}
\name{ISTDList}
\alias{ISTDList}
\alias{ISTDListPos}
\alias{ISTDListNeg}
\title{Example internal standard lists}
\format{
A \code{data.frame} with column \code{name}, \code{formula} and \code{rt}, specifying the analyte name,
  formula and retention time (in seconds), respectively.




}
\usage{
ISTDListPos

ISTDListNeg
}
\description{
The \code{ISTDListPos} and \code{ISTDListNeg} datasets are example lists with internal standards for a set of
environmental contaminants (measured in positive and negative ionization, respectively). The datasets can be directly
used with the \code{\link[patRoon]{normInts}} function.
}
\seealso{
\link{suspects}
}
\author{
Dominique Narain-Ford, Samira Absalah, Rick Helmus and Vittorio Albergamo
}
\keyword{datasets}
